refactor: Refactor rfcindex creation logic
